+++ src/ds/tnode_types/tnode_location.hpp
@@ -10,29 +10,29 @@ namespace treap_types {
 
 class LocationTnode : public Tnode {
   public:
-    uint32_t total_versions_in_subtree;
-    uint32_t total_nodes_in_subtree;
+    uint64_t total_versions_in_subtree;
+    uint64_t total_nodes_in_subtree;
 
-    LocationTnode(const uint32_t rc_total_versions,
-                  const uint32_t rc_total_nodes, 
-                  const uint32_t rc_data_id, 
+    LocationTnode(const uint64_t rc_total_versions,
+                  const uint64_t rc_total_nodes, 
+                  const uint64_t rc_data_id, 
                   const uint64_t rc_prio, 
-                  const uint32_t rc_index_tnode = 0xffffffff);
+                  const uint64_t rc_index_tnode = 0xffffffff);
     LocationTnode(const LocationTnode *oth);
-    LocationTnode(const uint32_t index);
+    LocationTnode(const uint64_t index);
 
-    static std::vector<uint32_t> saved_total_versions_subtree;
-    static std::vector<uint32_t> saved_total_nodes_subtree;
+    static std::vector<uint64_t> saved_total_versions_subtree;
+    static std::vector<uint64_t> saved_total_nodes_subtree;
 
-    static uint32_t next_h5_bp_index;
-    static Tnode* get_h5_tnode(const uint32_t req_data_id, const unsigned long long req_prio, const uint32_t req_index);
+    static uint64_t next_h5_bp_index;
+    static Tnode* get_h5_tnode(const uint64_t req_data_id, const unsigned long long req_prio, const uint64_t req_index);
     static void reset_get_h5_tnode(const H5::Group &group);
 
     static void append_tnode_data(Tnode *tnode);
     static void reset_append_tnode_data();
     static void write_tnode_data_to_h5(H5::Group &group);
 
-    static Tnode* create_new_specialized_tnode(const uint32_t data_id);
+    static Tnode* create_new_specialized_tnode(const uint64_t data_id);
     static Tnode* copy_specialized_tnode(const Tnode* oth);
 };
