+++ src/queries/count_indels.hpp
@@ -16,11 +16,11 @@ class CountIndelsQuery : public BaseQuery {
     bool before_lca_tnode = true;
 
   public:
-    std::vector<uint32_t> total_inserted_sequences;
-    std::vector<uint32_t> total_modified_sequences;
-    std::vector<uint32_t> total_deleted_sequences;
+    std::vector<uint64_t> total_inserted_sequences;
+    std::vector<uint64_t> total_modified_sequences;
+    std::vector<uint64_t> total_deleted_sequences;
     
-    CountIndelsQuery(const uint32_t num_total_snapshots);
+    CountIndelsQuery(const uint64_t num_total_snapshots);
     std::string get_treap_name();
     TreeDirectionToGo first_enter_into_node(const std::string &target_location_prefix, Tnode *, const BaseSortedTreap *, const ds::DB *);
     TreeDirectionToGo second_enter_into_node(const std::string &target_location_prefix, Tnode *, const BaseSortedTreap *, const ds::DB *);
